Remove legacy external Tor key generation
In the past, we didn't create the Tor onion keys ourselves. Back then,
we let Tor create the keys and persist those in Bisq.
