+++ network/network/src/main/java/bisq/network/p2p/node/transport/TorTransportService.java
@@ -7,7 +7,6 @@ import bisq.network.common.TransportType;
 import bisq.network.identity.NetworkId;
 import bisq.network.p2p.node.ConnectionException;
 import bisq.security.keys.KeyBundle;
-import bisq.security.keys.TorKeyGeneration;
 import bisq.tor.TorService;
 import bisq.tor.TorTransportConfig;
 import bisq.tor.onionservice.CreateOnionServiceResponse;
@@ -82,8 +81,7 @@ public class TorTransportService implements TransportService {
             bootstrapInfo.getBootstrapProgress().set(0.25);
             bootstrapInfo.getBootstrapDetails().set("Create Onion service for node ID '" + networkId + "'");
 
-            String privateOpenSshKey = TorKeyGeneration.getPrivateKeyInOpenSshFormat(keyBundle.getTorKeyPair().getPrivateKey());
-            CreateOnionServiceResponse response = torService.createOnionService(port, privateOpenSshKey, keyBundle.getTorKeyPair().getOnionAddress()).get(2, TimeUnit.MINUTES);
+            CreateOnionServiceResponse response = torService.createOnionService(port, keyBundle.getTorKeyPair(), keyBundle.getTorKeyPair().getOnionAddress()).get(2, TimeUnit.MINUTES);
 
             bootstrapInfo.getBootstrapState().set(BootstrapState.SERVICE_PUBLISHED);
             bootstrapInfo.getBootstrapProgress().set(0.5);
